some proof-of-concept work on a custom annotations layer
This commits adds a layer to the PDF where annotations outside of (but
on top of) the document can be drawn. It adds a highlight tool, which
lets you select text which then adds a highlight annotation to the new
annotation layer.

The annotation layer is scaled and rotated with the underlying page, so
things can be drawn in it at native PDF resolution.